refactor(configuracao): repository id integer

+++ src/main/java/com/corleone/configuracao/repository/ConfiguracaoRepository.java
@@ -3,5 +3,5 @@ package com.corleone.configuracao.repository;
 import com.corleone.configuracao.entity.Configuracao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConfiguracaoRepository extends JpaRepository<Configuracao, Long> {
+public interface ConfiguracaoRepository extends JpaRepository<Configuracao, Integer> {
 }
